+++ src/syrag/integrations/__init__.py
@@ -22,3 +22,11 @@ else:
             "create_syrag_query_tool",
         ]
     )
+
+try:
+    from syrag.integrations.llamaindex import LlamaIndexNodeChunker
+except ModuleNotFoundError:
+    pass
+else:
+    globals()["LlamaIndexNodeChunker"] = LlamaIndexNodeChunker
+    __all__.extend(["LlamaIndexNodeChunker"])